static microservice (#5)
* Added Static microservice routing configuration.

Now, for the static content, a microservice is invoked and it takes
care of everything. Also, this microservice is already containerized
which makes it easier to just deploy it.

Also, the apiGateway:v2Static image is created to show this change
instead of v1.

* Added more information about associated projects

API-Gateway is a gateway to the other microservices in its
eco-system. Updated more information about those.